refactoring :)
